Refactor ManagedList and Cart components to enhance item management and controls
- Updated ManagedList to accept item control components for better customization of item actions.
- Refactored Cart component to implement new item control actions for increasing, decreasing, and deleting cart items.
- Improved styling for item controls in managed lists to enhance user interaction.
- Adjusted utility functions for cart item management to ensure proper handling of active carts.

+++ portal/src/utilities/cart.ts
@@ -3,26 +3,44 @@ import { cartDAO, cartItemDAO } from "../DAO/cartDAO";
 import { getUser } from "./authentication";
 import type { Id } from "../types/common";
 
-export const fetchCartItems = async () => {
+export const getActiveCart = async () => {
     const activeCarts = await fetchActiveCarts();
     if (activeCarts && activeCarts.length > 0) {
-        if (activeCarts[0]) {
-            return await cartItemDAO.queryItems();
-        }
+        return activeCarts[0];
+    }
+    return null;
+}
+
+export const fetchCartItems = async () => {
+    const activeCart = await getActiveCart();
+    if (activeCart) {
+        return await cartItemDAO.queryItems();
     }
     return [];
 };
 
 export const postCartItem = async (cartItem: CartItemBase) => {
-    return await cartItemDAO.addItem(cartItem);
+    const activeCart = await getActiveCart();
+    if (activeCart) {
+        return await cartItemDAO.addItem({ ...cartItem, cart: activeCart.id });
+    }
+    return null;
 };
 
 export const deleteCartItem = async (id: Id) => {
-    return await cartItemDAO.removeItem(id);
+    const activeCart = await getActiveCart();
+    if (activeCart) {
+        return await cartItemDAO.removeItem(id);
+    }
+    return null;
 };
 
 export const updateCartItem = async (id: Id, updatedCartItem: CartItemBase) => {
-    return await cartItemDAO.amendItem(id, updatedCartItem);
+    const activeCart = await getActiveCart();
+    if (activeCart) {
+        return await cartItemDAO.amendItem(id, { ...updatedCartItem, cart: activeCart.id });
+    }
+    return null;
 };
